added unit tests for unknown stream and response path
